Convert analyzer.py into modules
