+++ emojifi_site/emojifi/emojifiers/clap.py
@@ -1,0 +1,9 @@
+import emoji
+
+
+def clappify_text(text):
+    """ Emojifies text by inserting claps around every word """
+    if text:
+        clap = emoji.emojize(' :clap: ', use_aliases=True)
+        return clap.join(text.split()) + clap
+    return ""
